Track ConnectionStates by unique Int keys
This change ensures that a connectionstate is uniquely
tied to a particular network connection. Reusing the
same network name when reconnecting will therefore
not cause messages in the event queue from the old
network to then apply to the new one. This chance also
eliminates a race condition with the network connection
exit reporting where aborting a connection that was
nearly finished could have aborted the exit reporting