Start introducing StateT
Here is an example of how the code would look like if it was translated into MonadState form. I'm not sure whether it's representative, but this is the first piece of code I saw that translates state in this package. Should this be done for the rest of the codebase?