Tweak command interpolation and expose disconnect time
Normal commands previously weren't interpolated. Now they are, but only
variables are exposed, not arguments like for macros.

The disconnect time is available as a variable during reconnecting.

A new variable form has been added to: ${variable|literal}
This denotes the value of the variable unless it is undefined, in which
case the literal value is used instead.