Allow changing padding at runtime
Now time values are split off from the image prefix
This means that I can dynamically change the amount of
Padding that should be rendered on each line.

Additionally this change uses a packed representation for
message timestamps. While this loses the pico-second time
Accuracy previously enjoyed, it means that the whole time
Value can be stored in 64-bit packed into the window-line