Sanitization of filepath for DCC transfers
At parse time, we remove path separators from the name and present just
that. When the user issues a "/dcc (accept|resume)" we check the
download directory for a file of the same name, if it is a directory we
fail, otherwise we assume that if the user said "resume" we want to
append to this file and if he said "accept" we overwrite.