Add /in command
This allows running commands like setwindow or mode
in contexts other than the currently focused window.

Also adds two new command types, adds focus to the context Args can work with,
and reworks a handful of commands to not fetch clientFocus themselves.
It also tweaks /list to not use extensionArg anymore.

There is still some work that needs doing in support of this one command
(as well as future metacommands but never mind):
- Tab completion needs to be burned to the ground.
- Views (names, masks) need to be changed to respect the override.
- `/in #channel-im-not-in masks b` should really just work.

ClientCommands don't make sense with /in and can exhibit surprising behavior,
and so /in currently rejects them.